Run a Demonstration of the r5rgui Application

A simple wrapper function to launch the r5rgui Shiny application using the sample Porto Alegre dataset included with the r5r package.

This function is designed for ease of use and demonstration purposes. It automatically handles the setup of the r5r_network object and provides default map settings (center and zoom) appropriate for the sample data.

Usage

r5r_gui_demo()

Value

This function does not return a value; it launches a Shiny application.

Details

The function first locates the sample data within the r5r package, then calls r5r::setup_r5() to build the routing core. Finally, it launches the Shiny app by calling r5r_gui() with these pre-configured objects. Please note that r5r may need to download the sample data on first use, which requires an internet connection.

Examples

if (FALSE) { # \dontrun{
# To run the demo application, simply call the function:
r5r_gui_demo()
} # }